refactor(bilibili): split bilibiliVideo into focused modules
Break the 1145-line bilibili/bilibiliVideo.ts into focused modules:
bilibiliVideoHelpers (BilibiliDownloaderHelper + subtitle/thumbnail path
helpers + yt-dlp failure redaction/formatting), bilibiliCoreDownload
(downloadVideo), and bilibiliSinglePart (downloadSinglePart).

bilibiliVideo.ts is now a barrel re-exporting downloadVideo and
downloadSinglePart. Splitting the two — where downloadSinglePart calls
downloadVideo — into separate modules with a one-way dependency
(singlePart -> coreDownload) keeps the import graph acyclic; the barrel
preserves the public surface for BilibiliDownloader, bilibiliCollection,
and the tests. Behavior is unchanged, including the issue #295 comments.
